refactor: 発話データを性格ごとにファイル分割
speechData.ts と contextualSpeechData.ts を
それぞれディレクトリに分割し、性格別ファイル + index.ts の構成に変更。
インポートパスは index.ts への自動解決により変更不要。

Closes #599
