--- src/game/speechData.ts
@@ -1,1334 +1,0 @@
-/**
- * キャラクター発話データ辞書
- * 性格タイプ × イベント種別ごとの発話バリエーションを定義
- */
-
-import type { Personality, SpeechEventType } from "../types";
-
-export const SPEECH_VARIANTS: Record<
-	Personality,
-	Record<SpeechEventType, readonly string[]>
-> = {
-	brave: {
-		move_success: [
-			"よし、進もう",
-			"この先は…",
-			"恐れるな、前へ！",
-			"道は俺が切り拓く",
-			"立ち止まるな！",
-			"迷う必要はない",
-			"俺について来い",
-			"一歩一歩が前進だ",
-			"この先に勝利がある",
-			"足を止めるな！",
-			"さあ、突き進め",
-			"後ろは振り返らない",
-			"行くぞ、仲間たち",
-			"剣を握る手に力が入る",
-			"この体で道を切り拓くぞ",
-		],
-		move_fail: [
-			"行き止まりか…",
-			"ここは通れない",
-			"別の道を探そう",
-			"壁か、ならば回り込む",
-			"引き返すのも勇気だ",
-			"壁一枚で諦めるか！",
-			"くっ、ここは塞がれている",
-			"この壁、剣で斬れないか",
-			"迂回路があるはずだ",
-			"焦るな、道は必ずある",
-			"壁は俺を止められない",
-			"ここじゃない、別の道だ",
-			"力ずくで壊したいところだが",
-			"退くのは負けじゃない",
-			"別の突破口を見つける",
-		],
-		attack_miss: [
-			"空振りだ！",
-			"くっ、外した",
-			"手応えがない…",
-			"次は必ず当てる！",
-			"まだだ、諦めない！",
-			"甘かったか…！",
-			"この程度で終わらん！",
-			"剣筋が甘かった",
-			"もう一撃、叩き込む！",
-			"読まれたか…だが！",
-			"反省は後だ、もう一度！",
-			"焦ったか…冷静にいけ",
-			"敵もやるな…！",
-			"この腕で仕留めてみせる！",
-			"外したが、次は違う！",
-		],
-		combo_activated: [
-			"今だ！畳みかける！",
-			"チャンスだ！",
-			"連続攻撃！",
-			"止まるなよ！",
-			"一気に決める！",
-			"怒涛の連撃だ！",
-			"この勢いで押し切る！",
-			"逃がさないぞ！",
-			"俺の剣が唸る！",
-			"休む暇は与えない！",
-			"全力で叩き込め！",
-			"ここで仕留める！",
-			"連撃で決着だ！",
-			"勝負はここだ！",
-			"剣の舞だ、食らえ！",
-		],
-		enemy_defeated: [
-			"やった！",
-			"倒したぞ",
-			"まだ先がある",
-			"俺の勝ちだ！",
-			"強敵だったな",
-			"これで道が開けた",
-			"次の敵も来い！",
-			"まだまだ行けるぞ",
-			"この剣に斬れぬものはない",
-			"この力、信じてよかった",
-			"一つ片付いたな",
-			"手加減はしない主義だ",
-			"見事な戦いだった",
-			"この体格で押し切ったぞ",
-			"さらに強くなれる",
-		],
-		damage_taken: [
-			"くっ…！",
-			"痛い！",
-			"まだ耐えられる",
-			"この程度で倒れるか！",
-			"効いたが…まだだ！",
-			"かすり傷だ！",
-			"俺はまだ立てる！",
-			"この体はそう簡単には折れん！",
-			"歯を食いしばれ…！",
-			"ここで倒れてたまるか",
-			"傷が増えるほど強くなる！",
-			"痛みは戦士の勲章だ",
-			"やるな…だが！",
-			"膝はまだ折れていない！",
-			"根性で耐えてみせる！",
-		],
-		game_over: [
-			"ここまでか…",
-			"無念…",
-			"次こそは…",
-			"悔しい…だが良い戦いだった",
-			"俺はまだ…終わらない…",
-			"力が…足りなかった",
-			"すまない…仲間たち",
-			"この敗北を糧にする",
-			"まだ倒れるわけには…",
-			"もっと強くならねば",
-			"剣を握る力がもう…",
-			"次は…必ず勝つ",
-			"この悔しさ、忘れない",
-			"立ち上がれ…俺",
-			"この体が動く限り、戦いは終わらない…",
-		],
-		trap_triggered: [
-			"しまった、罠だ！",
-			"うわっ！",
-			"不覚…！",
-			"くそ、引っかかった！",
-			"罠ごときに…！",
-			"油断した…！",
-			"こんなもので俺を！",
-			"足元を見てなかった…！",
-			"罠師め…やるな",
-			"次は踏まんぞ！",
-			"この屈辱は忘れん！",
-			"卑怯な手を…！",
-			"ぐっ…この体格じゃ避けにくいな",
-			"剣で斬れる罠ならいいのに",
-			"二度目はないぞ！",
-		],
-		treasure_found: [
-			"お、宝箱だ！",
-			"ありがたい",
-			"回復できた",
-			"よし、これで態勢を立て直せる",
-			"運も実力のうちだ！",
-			"天が俺に味方した！",
-			"これで戦える！",
-			"戦士への褒美だな",
-			"助かる、使わせてもらう",
-			"この先の戦いに備えよう",
-			"よし、まだやれるぞ",
-			"思わぬ収穫だ",
-			"剣を振るう力が戻った",
-			"これがあれば心強い",
-			"戦いの神に感謝だ",
-		],
-		rest_area_used: [
-			"少し休もう",
-			"体力が戻った",
-			"ふぅ、一息つける",
-			"休息も戦いの一部だ",
-			"英気を養うぞ",
-			"ここで力を蓄える",
-			"体が軽くなった",
-			"束の間の休息だな",
-			"剣の手入れもしておくか",
-			"次の戦いに備える",
-			"安らげる場所があるとは",
-			"この体を万全にしておくぞ",
-			"万全の状態で挑む",
-			"戦士にも休息は必要だ",
-			"よし、準備完了だ",
-		],
-		floor_reached: [
-			"新しい階層だ",
-			"ここからが本番",
-			"気を引き締めよう",
-			"さあ、行くぞ！",
-			"どんな敵が待っていようと！",
-			"新たな戦場だ！",
-			"さらなる試練か、望むところだ",
-			"この先には何がある",
-			"俺の剣が試される時だ",
-			"強敵よ、かかって来い！",
-			"階層が変わっても俺は変わらん",
-			"深く潜るほど燃えてくる",
-			"ここで引き返す男じゃない",
-			"新たな冒険の始まりだ",
-			"覚悟はできている",
-		],
-		jump_success: [
-			"突っ込むぞ！",
-			"一気に行く！",
-			"ここを抜ける！",
-			"飛び込め！",
-			"勢いで突破だ！",
-			"飛んだ！着地！",
-			"大胆に行くぞ！",
-			"この体重で跳べるとはな！",
-			"勢いは止まらない！",
-			"華麗に決めた！",
-			"飛び越えてやったぞ！",
-			"この程度、朝飯前だ",
-			"勇気ある一歩だ！",
-			"躊躇なく飛ぶ！",
-			"行ける、行けるぞ！",
-		],
-		card_acquired: [
-			"新しい力だ！",
-			"これは使えるぞ！",
-			"戦力強化だ！",
-			"この手札で戦い抜く！",
-			"頼もしい一枚だ",
-			"よし、手に入れたぞ！",
-			"これで戦い方が広がる！",
-			"強くなれる気がする！",
-			"俺の手札に加わるか！",
-			"この力、受け取った！",
-			"新しい武器だな！",
-			"デッキが強化された！",
-			"いい獲物だ！",
-			"この一枚で勝ちに行く！",
-			"戦いの幅が広がるぞ！",
-		],
-		card_skipped: [
-			"今の手札で十分だ",
-			"必要ないな",
-			"俺にはこの手札がある",
-			"見送るぞ",
-			"今のデッキを信じる",
-			"手札の完成度は十分だ",
-			"これ以上は不要だ",
-			"今の構成で戦い抜く",
-			"欲張りすぎは禁物だ",
-			"この手札が俺の武器だ",
-			"変える必要はない",
-			"温存しておく",
-			"今は見送りだ",
-			"現状維持で十分強い",
-			"このデッキで突き進む！",
-		],
-	},
-	cautious: {
-		move_success: [
-			"周囲を確認…問題ない",
-			"一歩ずつ、慎重に",
-			"罠がないか注意しよう",
-			"安全を確認して進む",
-			"この方向で合っているはず",
-			"地形に変化はないな",
-			"視界良好、進行可能",
-			"足元を確認…安全だ",
-			"次の一手を考えながら進む",
-			"予定ルート通りだ",
-			"周辺の敵影なし、前進する",
-			"リスクは最小限に抑えた",
-			"眼鏡を直しつつ…地形を確認",
-			"想定内の道だ",
-			"慎重に、だが確実に",
-		],
-		move_fail: [
-			"予想通り、通れないな",
-			"別ルートを検討しよう",
-			"ここは危険かもしれない",
-			"想定内だ、迂回する",
-			"事前に確認しておくべきだった",
-			"代替ルートに切り替える",
-			"壁の構造を記憶しておこう",
-			"想定外だが対処可能だ",
-			"ここは通行不可、記録した",
-			"マップの修正が必要だ",
-			"予備プランに移行する",
-			"眼鏡越しに壁を観察…素材は石か",
-			"情報が足りなかったか",
-			"安全な迂回路を探す",
-			"壁の位置を覚えておく",
-		],
-		attack_miss: [
-			"読みが外れた…",
-			"距離を見誤ったか",
-			"次は確実に仕留める",
-			"焦りは禁物だ",
-			"冷静に、もう一度",
-			"敵の動きを再分析する",
-			"角度の計算を修正しよう",
-			"パターンを読み直す",
-			"誤差の原因を特定する",
-			"想定と実際のずれか",
-			"眼鏡がずれた…修正して再試行",
-			"この失敗から学べることがある",
-			"感情に流されるな",
-			"集中力を維持しろ",
-			"次こそ精度を上げる",
-		],
-		combo_activated: [
-			"計算通りだ",
-			"読み通りの展開",
-			"ここが好機、畳みかける",
-			"想定した連携だ",
-			"隙を見逃さない",
-			"シミュレーション通りの結果だ",
-			"予測していた展開だ",
-			"ここで押し切るのが最善",
-			"効率的に攻撃を重ねる",
-			"全て計画の範囲内だ",
-			"リスクを最小化しつつ攻める",
-			"分析が功を奏した",
-			"最適解で攻め続ける",
-			"冷静に連撃を叩き込む",
-			"眼鏡の奥で勝機を捉えた",
-		],
-		enemy_defeated: [
-			"予定通り排除した",
-			"脅威を一つ除去",
-			"油断はまだ早い",
-			"次の敵に備えよう",
-			"分析通りの弱点だった",
-			"被害を最小限に抑えられた",
-			"想定時間内に排除完了",
-			"このパターンは記憶した",
-			"次はもっと効率よく倒せる",
-			"周囲に残存敵がいないか確認",
-			"データを蓄積しておこう",
-			"勝因を分析しておく",
-			"予備の戦術は不要だったな",
-			"計画通りの戦闘結果だ",
-			"冷静な判断が勝利を呼んだ",
-		],
-		damage_taken: [
-			"想定よりダメージが大きい",
-			"防御が甘かった…",
-			"回復手段を確認しないと",
-			"ここは退くべきか…",
-			"被害を最小限に抑えねば",
-			"被弾原因を分析する",
-			"体力の残量を確認…",
-			"この攻撃パターンを記憶した",
-			"防御態勢を見直す必要がある",
-			"リスク評価を修正しなければ",
-			"眼鏡が…割れてなければいいが",
-			"ダメージ量は許容範囲内か…",
-			"痛手だが致命的ではない",
-			"立て直しの余地はまだある",
-			"撤退ラインを再設定する",
-		],
-		game_over: [
-			"判断を誤ったか…",
-			"もっと慎重にすべきだった",
-			"データは次に活かせる",
-			"敗因を分析しなければ…",
-			"この経験は無駄にしない",
-			"リスク評価が甘かった…",
-			"見落としがあったか…",
-			"より精密な計画が必要だった",
-			"仮説が間違っていた…",
-			"データ不足が敗因だ",
-			"撤退タイミングを逸した…",
-			"冷静さを欠いたか…",
-			"次回はこの失敗を糧にする",
-			"分析し直さなければ…",
-			"全てのパターンを想定できなかった",
-		],
-		trap_triggered: [
-			"見落としていた…不覚",
-			"罠の兆候を見逃したか",
-			"次からはもっと注意を",
-			"確認が甘かった…",
-			"この種の罠は記憶した",
-			"チェック手順に漏れがあった",
-			"この罠のパターンを記録する",
-			"注意力が散漫だったか…",
-			"安全確認を怠った結果だ",
-			"同種の罠を警戒リストに追加",
-			"眼鏡を拭いていたら踏んだ…",
-			"もう一段階慎重にすべきだった",
-			"この失敗は二度と繰り返さない",
-			"予想外の罠の設置位置だ",
-			"探知方法を改善しなければ",
-		],
-		treasure_found: [
-			"回復アイテムか、助かる",
-			"今の状態なら使い時だ",
-			"貴重な回復の機会だ",
-			"ありがたい、備蓄しておきたいが",
-			"これで少し余裕ができた",
-			"状況改善の好機だ",
-			"使用タイミングを見極めよう",
-			"予備のリソースが増えた",
-			"これは戦略的に有利になる",
-			"残り体力と相談して使おう",
-			"計画の修正が必要かもしれない",
-			"ここに宝箱がある確率は考慮済み",
-			"有効活用できるよう計画しよう",
-			"危機的状況の回避に役立つ",
-			"回復分を計算に組み込もう",
-		],
-		rest_area_used: [
-			"ここで態勢を整えよう",
-			"安全を確認してから休む",
-			"回復しつつ次の計画を練る",
-			"体力管理は最重要だ",
-			"焦らず回復を優先する",
-			"休息の間にデータを整理する",
-			"次の行動計画を立てよう",
-			"ここが安全地帯なら最大限活用する",
-			"体力と精神力を回復させる",
-			"先のルートを検討しておこう",
-			"冷静になれる貴重な時間だ",
-			"眼鏡を磨きつつ情報を整理する",
-			"体調管理も戦略の一部だ",
-			"しっかり休んで万全を期す",
-			"回復後の行動方針を固めておく",
-		],
-		floor_reached: [
-			"新しい階層か、偵察から始めよう",
-			"まず地形を把握する",
-			"未知の敵に警戒を",
-			"情報収集が先だ",
-			"慎重に進めていこう",
-			"新たなデータ収集の機会だ",
-			"前の階層の知見を活かそう",
-			"敵の配置パターンを予測する",
-			"安全な拠点を確保したい",
-			"環境の変化に適応しなければ",
-			"警戒レベルを上げておく",
-			"先入観を排除して観察しよう",
-			"眼鏡を直して…新たな分析開始",
-			"リスク要因を洗い出そう",
-			"あらゆる可能性に備えておく",
-		],
-		jump_success: [
-			"計算通り、着地成功",
-			"リスクはあったが通過できた",
-			"飛距離は想定内だ",
-			"着地点の安全を確認",
-			"大胆な手だが必要だった",
-			"着地の衝撃は許容範囲内",
-			"予測座標に正確に着地",
-			"この手段が最善だった",
-			"想定通りの軌道だ",
-			"安全マージンは確保できた",
-			"着地後の周囲を確認する",
-			"計算を信じて飛んだ甲斐があった",
-			"リスクとリターンの判断は正しかった",
-			"次も同じ判断ができるように記録する",
-			"飛距離のデータを記録しておこう",
-		],
-		card_acquired: [
-			"カードの性能を確認しよう",
-			"デッキの最適化が可能だ",
-			"戦力の見直しが必要だ",
-			"この一枚の価値を分析する",
-			"手札構成の改善に使えるか",
-			"有効な選択肢が増えた",
-			"デッキバランスを考慮しよう",
-			"期待値が上がる組み合わせだ",
-			"この追加は計画通りだ",
-			"戦術の幅が広がる",
-			"慎重に入れ替えを検討する",
-			"リスクとリターンを天秤にかけよう",
-			"シナジーのある構成を目指す",
-			"この一枚でデッキが完成に近づく",
-			"眼鏡越しにカードを精査する",
-		],
-		card_skipped: [
-			"現状のデッキが最適解だ",
-			"入れ替えるメリットが薄い",
-			"無理に変更する必要はない",
-			"リスクを増やすべきではない",
-			"今の構成が安定している",
-			"不要な変更は避ける",
-			"現状維持が最善の判断だ",
-			"デッキの安定性を優先する",
-			"変更による損失リスクが高い",
-			"見送りが合理的な選択だ",
-			"この判断は間違っていない",
-			"分析の結果、不要と判断した",
-			"今のバランスを崩したくない",
-			"次の機会を待とう",
-			"冷静な判断ができた",
-		],
-	},
-	cheerful: {
-		move_success: [
-			"さくさく進むよ〜！",
-			"いい感じ！",
-			"次はどこかな〜",
-			"わくわくするね！",
-			"どんどん行こう！",
-			"ルンルン♪",
-			"お散歩気分〜！",
-			"今日はいい日だな〜",
-			"この道、好きかも！",
-			"快調快調〜！",
-			"冒険って楽しい！",
-			"小さい体でもどんどん進むよ！",
-			"足取り軽いよ〜",
-			"狭い隙間もスイスイだね！",
-			"たのしいな〜！",
-		],
-		move_fail: [
-			"あれ〜、行き止まり！",
-			"んー、こっちじゃないか",
-			"まあ、別の道もあるし！",
-			"ドンマイドンマイ！",
-			"あはは、戻ろう！",
-			"おっとっと、壁だ〜",
-			"行き止まりも冒険のうち！",
-			"残念〜、でも大丈夫！",
-			"違う道にも何かあるかも！",
-			"ここは壁だね〜",
-			"まあまあ、気にしない！",
-			"回り道も楽しいよ！",
-			"はずれだ〜、次いこ次！",
-			"壁さんこんにちは〜！",
-			"小さいから隙間ないか探してみよ〜",
-		],
-		attack_miss: [
-			"あちゃー、外しちゃった！",
-			"惜しい〜！",
-			"次は当てるぞ〜！",
-			"えへへ、ごめんごめん",
-			"もう一回！",
-			"あらら〜、空振り！",
-			"ドンマイ自分！",
-			"惜しかったのに〜！",
-			"次こそは！えいっ！",
-			"スカッと空振り〜！",
-			"あたらなかった〜",
-			"リーチが足りないかな〜！",
-			"えーい、待ってよ〜！",
-			"がんばれ自分！",
-			"外しちゃったけど元気！",
-		],
-		combo_activated: [
-			"やったやった！連続だ！",
-			"ノリノリだね！",
-			"すごいすごい！",
-			"いっけー！",
-			"コンボだ〜！楽しい！",
-			"テンション上がる〜！",
-			"絶好調だよ〜！",
-			"どんどんいくよ〜！",
-			"連続ヒット〜！最高！",
-			"止まらないよ〜！",
-			"コンボって気持ちいい！",
-			"きたきたきた〜！",
-			"小さくても強いんだから〜！",
-			"もっともっと！",
-			"この感じ、サイコー！",
-		],
-		enemy_defeated: [
-			"やったね！",
-			"勝った勝った〜！",
-			"ナイスファイト！",
-			"いえーい！",
-			"よっしゃー！",
-			"お疲れさまでした〜！",
-			"勝っちゃった〜！",
-			"ばんざーい！",
-			"楽しかった〜！",
-			"すっきり〜！",
-			"また勝っちゃった！",
-			"小さい体でもやるでしょ！",
-			"大勝利〜！",
-			"ナイスナイス！",
-			"これで安心だね！",
-		],
-		damage_taken: [
-			"いたた！",
-			"うぅ、やるね〜",
-			"まだまだ元気だよ！",
-			"ちょっと痛かった…えへへ",
-			"大丈夫大丈夫！",
-			"あいたた〜！",
-			"やられた〜、でも平気！",
-			"ちくっとしたかな？",
-			"負けないもんね〜！",
-			"痛いけどへっちゃら！",
-			"ぐぬぬ…でも笑顔！",
-			"小さい体にはちょっと響くかも〜",
-			"これくらいなんともない！",
-			"ちょっとびっくりしただけ！",
-			"痛いの痛いの飛んでけ〜！",
-		],
-		game_over: [
-			"あちゃ〜、やられちゃった",
-			"また挑戦しよう！",
-			"次はきっとうまくいくよ！",
-			"楽しかった〜！もう一回！",
-			"ドンマイ！切り替えていこう",
-			"おしかったな〜",
-			"次こそ頑張る〜！",
-			"いい勉強になったね！",
-			"まだまだこれから！",
-			"失敗は成功のもとだよ！",
-			"今度はもっと上手にやるぞ〜",
-			"やられちゃったけど楽しかった！",
-			"ま、こういう日もあるよね！",
-			"またすぐ戻ってくるよ〜！",
-			"リベンジだ〜！",
-		],
-		trap_triggered: [
-			"きゃっ！罠だった！",
-			"びっくりした〜！",
-			"わわっ、やられた！",
-			"あはは、引っかかっちゃった",
-			"もう〜、意地悪な罠！",
-			"うわわ〜！罠だ〜！",
-			"ひゃっ！びっくり！",
-			"やだやだ〜、罠だなんて！",
-			"あちゃ〜、踏んじゃった",
-			"いたっ！もう〜！",
-			"サプライズは嫌いだよ〜",
-			"小さいから見えなかったよ〜",
-			"次は気をつけるね〜！",
-			"えーん、罠にやられた〜",
-			"ドッキリかと思った〜！",
-		],
-		treasure_found: [
-			"わーい、宝箱だ！",
-			"ラッキー！",
-			"やった〜、うれしい！",
-			"ご褒美だね！",
-			"いいもの見つけた〜！",
-			"宝箱だ〜！開けちゃお！",
-			"わくわく、何が入ってるかな！",
-			"きらきら〜！",
-			"プレゼントみたい！",
-			"うれしいうれしい！",
-			"今日はついてるね〜！",
-			"ハッピー！",
-			"これはうれしいサプライズ！",
-			"お宝ゲット〜！",
-			"大当たり〜！",
-		],
-		rest_area_used: [
-			"ちょっと休憩〜！",
-			"はぁ〜、気持ちいい！",
-			"元気回復！",
-			"リフレッシュ完了！",
-			"よーし、また頑張ろ！",
-			"のんびりタイム〜！",
-			"ほっと一息〜！",
-			"疲れが吹き飛んだ〜！",
-			"ごろんってしたいな〜",
-			"ぽかぽかする〜",
-			"充電完了！",
-			"小さい体はすぐ回復するよ〜！",
-			"休憩って最高！",
-			"体が軽くなった〜！",
-			"エネルギー満タン！",
-		],
-		floor_reached: [
-			"新しいフロアだ〜！",
-			"わくわく！何があるかな",
-			"楽しくなってきた！",
-			"どんどん奥へ！",
-			"冒険の続きだね！",
-			"新しい景色だ〜！",
-			"探検探検〜！",
-			"ここはどんなところかな！",
-			"テンション上がる〜！",
-			"もっと深くまで行くよ！",
-			"新ステージだ〜！",
-			"きゃー、楽しみ！",
-			"ドキドキワクワク！",
-			"新しい出会いがあるかな〜",
-			"わーい、次のフロア！",
-		],
-		jump_success: [
-			"とりゃー！",
-			"ジャンプ成功！",
-			"飛んだ飛んだ〜！",
-			"えいっ！着地！",
-			"気持ちいい〜！",
-			"ぴょーん！",
-			"着地成功〜！",
-			"飛ぶの楽しい〜！",
-			"身軽でしょ〜！",
-			"ひゅーん！とんだ〜！",
-			"やった〜、飛べた！",
-			"ジャンプ大好き！",
-			"小さい体は軽いから楽チン！",
-			"すごい跳んだ〜！",
-			"ばっちり着地！",
-		],
-		card_acquired: [
-			"わ〜い、新しいカード！",
-			"どれと交換しようかな〜！",
-			"きらきらしてる〜！",
-			"いいの見つけた〜！",
-			"カードゲットだよ〜！",
-			"やった〜、新戦力！",
-			"これ強そう〜！",
-			"うれしいうれしい！",
-			"デッキが賑やかになるね！",
-			"どれにしよっかな〜！",
-			"わくわくする〜！",
-			"お宝カードだ〜！",
-			"ラッキー！いいカード！",
-			"これは使えそう〜！",
-			"新しいカードって楽しいね！",
-		],
-		card_skipped: [
-			"んー、今はいいかな〜！",
-			"またの機会にね〜！",
-			"今のデッキが好きだもん！",
-			"パスパス〜！",
-			"今回はいいや〜！",
-			"今の手札がお気に入り〜！",
-			"見送っちゃおう！",
-			"ドンマイ〜、次があるよ！",
-			"今は変えなくて大丈夫！",
-			"このままでいっこう〜！",
-			"スキップだよ〜！",
-			"今の手札で頑張る〜！",
-			"また次のチャンスがあるし！",
-			"今はこのままでオッケー！",
-			"お気に入りは変えないもん〜！",
-		],
-	},
-	stoic: {
-		move_success: [
-			"…前進",
-			"進む",
-			"問題ない",
-			"…",
-			"道は続く",
-			"…行ける",
-			"静かだ",
-			"…先へ",
-			"足音だけが響く",
-			"…進行",
-			"道がある",
-			"…歩く",
-			"前へ",
-			"…昔もこんな道を歩いた",
-			"何もない…良いことだ",
-		],
-		move_fail: [
-			"…壁か",
-			"通れない",
-			"別の道を",
-			"…行き止まり",
-			"戻る",
-			"…ここではない",
-			"迂回する",
-			"…無理だな",
-			"壁だ",
-			"…退く",
-			"道を変える",
-			"…仕方ない",
-			"別の方向へ",
-			"引き返す",
-			"…この体格では抜けられんか",
-		],
-		attack_miss: [
-			"…外した",
-			"甘い",
-			"次",
-			"…精度が足りない",
-			"もう一度",
-			"…まだだ",
-			"足りない",
-			"…修正する",
-			"当てる",
-			"…集中",
-			"甘かった",
-			"…次こそ",
-			"もう一撃",
-			"…戦場では許されん失態だ",
-			"外れた…だが",
-		],
-		combo_activated: [
-			"…畳みかける",
-			"今だ",
-			"続ける",
-			"隙あり",
-			"…連撃",
-			"…止めない",
-			"追撃",
-			"…ここで",
-			"攻める",
-			"…好機",
-			"逃さない",
-			"…押す",
-			"一気に",
-			"…歴戦の勘が告げている",
-			"手を緩めない",
-		],
-		enemy_defeated: [
-			"…終わりだ",
-			"排除完了",
-			"次",
-			"…片付いた",
-			"先へ",
-			"…倒した",
-			"終わった",
-			"…次の敵は",
-			"処理完了",
-			"…問題ない",
-			"片付けた",
-			"…先を急ぐ",
-			"一つ減った",
-			"…幾度目の勝利か、数えてもいない",
-			"次へ行く",
-		],
-		damage_taken: [
-			"…効くな",
-			"浅い",
-			"まだだ",
-			"…耐える",
-			"構わない",
-			"…痛みは知っている",
-			"この程度",
-			"…まだ立てる",
-			"問題ない",
-			"…致命傷ではない",
-			"耐えた",
-			"…平気だ",
-			"傷は浅い",
-			"…この体は幾多の傷を負ってきた",
-			"まだ戦える",
-		],
-		game_over: [
-			"…ここまでか",
-			"力及ばず",
-			"…無念",
-			"まだ…終われない",
-			"次は…負けない",
-			"…すまない",
-			"足りなかった",
-			"…認める",
-			"力が…",
-			"…これが限界か",
-			"不甲斐ない",
-			"…悔いはない",
-			"また、いつか",
-			"…力尽きた",
-			"…長い戦いだった",
-		],
-		trap_triggered: [
-			"…罠か",
-			"不覚",
-			"…やられた",
-			"浅はか",
-			"…注意が足りなかった",
-			"…失態",
-			"踏んだ",
-			"…気づかなかった",
-			"油断した",
-			"…迂闊",
-			"二度目はない",
-			"…侮った",
-			"甘かった",
-			"…戦場の罠はもっと巧妙だった",
-			"罠を踏むとは",
-		],
-		treasure_found: [
-			"…助かる",
-			"回復する",
-			"ありがたい",
-			"…使わせてもらう",
-			"丁度良い",
-			"…拾う",
-			"使える",
-			"…悪くない",
-			"これは良い",
-			"…必要なものだ",
-			"ありがたく",
-			"…助かった",
-			"いただく",
-			"…運が向いた",
-			"活用する",
-		],
-		rest_area_used: [
-			"…少し休む",
-			"回復",
-			"態勢を整える",
-			"…一息",
-			"ここで",
-			"…休息",
-			"体力を戻す",
-			"…しばし",
-			"休む",
-			"…助かる",
-			"体を休める",
-			"…古傷が疼く。少し休もう",
-			"少し、休む",
-			"…回復した",
-			"整えた",
-		],
-		floor_reached: [
-			"…新たな階層",
-			"進む",
-			"油断するな",
-			"…次の戦場",
-			"気を緩めるな",
-			"…降りる",
-			"新しい敵か",
-			"…構わない",
-			"どこまでも行く",
-			"…深くなる",
-			"先がある",
-			"…次の階だ",
-			"まだ続く",
-			"…昔もこんな深みに潜ったことがある",
-			"変わらず行く",
-		],
-		jump_success: [
-			"…飛ぶ",
-			"突破",
-			"一気に",
-			"…抜ける",
-			"行く",
-			"…跳んだ",
-			"着地",
-			"…問題ない",
-			"飛んだ",
-			"…それだけだ",
-			"越えた",
-			"…着いた",
-			"通過",
-			"…この体格で跳ぶのは得意ではないが",
-			"何でもない",
-		],
-		card_acquired: [
-			"…新しい手札",
-			"受け取る",
-			"使える",
-			"…悪くない",
-			"一枚追加",
-			"…確認した",
-			"手に入れた",
-			"…戦力になる",
-			"選ぶ",
-			"…これだ",
-			"加える",
-			"…必要なものだ",
-			"ありがたい",
-			"…デッキに入れる",
-			"判断する",
-		],
-		card_skipped: [
-			"…不要",
-			"見送る",
-			"今は要らない",
-			"…このままで",
-			"変えない",
-			"…十分だ",
-			"必要ない",
-			"…現状維持",
-			"見送った",
-			"…問題ない",
-			"変更なし",
-			"…今の手札で",
-			"不要だ",
-			"…パスする",
-			"このままでいい",
-		],
-	},
-	curious: {
-		move_success: [
-			"この先に何があるだろう",
-			"面白い構造だな",
-			"どんな発見があるかな",
-			"もっと奥へ行ってみよう",
-			"この通路は気になるな",
-			"壁の模様が独特だ",
-			"この道の先が気になる",
-			"新しい発見の予感がする",
-			"地形のパターンが見えてきた",
-			"ここの空気が変わった気がする",
-			"なぜこの通路はこの形なんだろう",
-			"手帳に地形を書き留めておこう",
-			"この建築様式は興味深い",
-			"まだまだ知らないことだらけだ",
-			"観察しながら進もう",
-		],
-		move_fail: [
-			"ここは行けないのか…残念",
-			"壁の向こうには何が？",
-			"別の道に何かあるかも",
-			"なぜここは塞がれてるんだろう",
-			"迂回路を探してみよう",
-			"この壁の材質が気になる",
-			"行き止まりにも意味があるはず",
-			"壁の厚さはどれくらいだろう",
-			"なぜここだけ通れないのか",
-			"壁の向こうの構造が知りたい",
-			"行き止まりか、標本採取でもしておこう",
-			"この配置にはパターンがあるかも",
-			"壁にも何かヒントがあるかもしれない",
-			"塞がれている理由が気になる",
-			"ふむ、ここは通れないのか",
-		],
-		attack_miss: [
-			"なぜ外れた？距離かな",
-			"この敵の回避パターンは…",
-			"どうすれば当たるだろう",
-			"興味深い動きをするな",
-			"次はタイミングを変えてみよう",
-			"この敵の反射速度は予想以上だ",
-			"攻撃の軌道を分析してみよう",
-			"なるほど、そう避けるのか",
-			"回避のメカニズムが気になる",
-			"面白い、別の角度から試そう",
-			"どの距離なら当たるだろう",
-			"この敵の回避能力は論文にできそうだ",
-			"外れた原因を突き止めたい",
-			"攻撃方法を変えてみようか",
-			"データが一つ増えた",
-		],
-		combo_activated: [
-			"この組み合わせが効くのか！",
-			"面白い！連鎖が起きた！",
-			"なるほど、こう繋がるのか",
-			"予想以上の効果だ！",
-			"もっと試してみたいな",
-			"この連携パターンは記録しておこう",
-			"コンボのメカニズムが見えた！",
-			"どこまで連鎖するか試したい",
-			"興味深い相乗効果だ",
-			"この発見は大きいぞ",
-			"もっと効率的な組み合わせがあるかも",
-			"連鎖の法則が分かってきた",
-			"素晴らしい、仮説が実証された",
-			"このパターン、応用できそうだ",
-			"手帳に書ききれないくらいの発見だ",
-		],
-		enemy_defeated: [
-			"この敵は何者だったんだろう",
-			"倒せた！弱点はあそこか",
-			"興味深い戦闘だった",
-			"もっと強い敵もいるのかな",
-			"どんな仕組みで動いてたんだろう",
-			"この敵の生態が気になる",
-			"戦闘データが蓄積されていく",
-			"次はもっと効率的に倒せるかも",
-			"この敵固有の特徴は何だったんだろう",
-			"貴重な観察の機会だった",
-			"倒してしまったか…もう少し観察したかった",
-			"戦闘パターンの分析が進んだ",
-			"似た敵と比較してみたいな",
-			"この敵から学べることは多い",
-			"調査ノートに詳しく記録しておこう",
-		],
-		damage_taken: [
-			"この攻撃はどういう原理だ？",
-			"痛いけど…学びがある",
-			"なるほど、そう来るか",
-			"この敵の攻撃パターンは…",
-			"次は避けられるかな",
-			"この痛みの原因を分析しよう",
-			"攻撃の威力と速度を計測したい",
-			"面白い攻撃方法だな",
-			"被弾データとして記録しておこう",
-			"なぜこのタイミングで攻撃が来た？",
-			"この攻撃の仕組みが気になる",
-			"痛いが、興味深いデータだ",
-			"防御の改善点が見えてきた",
-			"ローブが破れた…繕わないと",
-			"次は対策を立てられそうだ",
-		],
-		game_over: [
-			"まだ知りたいことがあったのに…",
-			"この先に何があったんだろう",
-			"もう少し調べたかった…",
-			"敗因は何だったんだろう",
-			"次はもっとうまくやれるはず",
-			"途中で終わるなんて残念だ",
-			"集めたデータは無駄にしない",
-			"まだ解明できていないことがある",
-			"この経験から何を学べるか…",
-			"もっと観察する時間が欲しかった",
-			"答えに辿り着けなかった…",
-			"次こそ最奥部まで行きたい",
-			"未知の領域がまだたくさんある",
-			"失敗の原因を究明しなければ",
-			"好奇心は尽きない…また挑もう",
-		],
-		trap_triggered: [
-			"これは…罠か！どんな仕組みだ？",
-			"巧妙な罠だ、誰が作ったんだろう",
-			"この罠の構造が気になる",
-			"なるほど、ここに仕掛けてあったか",
-			"次は見抜けるかもしれない",
-			"この罠の起動メカニズムは…",
-			"痛いけど、仕掛けは見事だ",
-			"どんな技術で作られたんだろう",
-			"罠の設計思想が興味深い",
-			"トリガーの位置を記憶した",
-			"同じ原理の罠が他にもあるかも",
-			"この罠、分解して調べたいな",
-			"予想外の場所に仕掛けてあった",
-			"ローブの裾が引っかかったみたい",
-			"これは新しい発見だ…痛いけど",
-		],
-		treasure_found: [
-			"おっ、何が入ってるかな",
-			"これは珍しい！",
-			"どこから来た宝なんだろう",
-			"中身が気になる！",
-			"こんなところに宝箱が…不思議だ",
-			"誰がここに置いたんだろう",
-			"宝箱の装飾が興味深い",
-			"中身を詳しく調べたいな",
-			"この宝箱の由来が気になる",
-			"なぜこの場所に宝があるのか",
-			"貴重な発見だ！",
-			"宝箱のロック機構が面白い",
-			"この素材はどこ産だろう",
-			"回復できる…仕組みが知りたい",
-			"宝箱にもいろいろ種類があるんだな",
-		],
-		rest_area_used: [
-			"ここはなぜ安全なんだろう",
-			"休みながら考えよう",
-			"この場所の成り立ちが気になる",
-			"回復しつつ次の探索を計画",
-			"不思議な空間だな…",
-			"なぜここだけ敵がいないんだろう",
-			"この場所の構造を観察しておこう",
-			"休息中に手帳を整理しよう",
-			"安全地帯の原理が気になる",
-			"ここの空気が他と違う気がする",
-			"回復の仕組みは何なんだろう",
-			"この場所にも秘密がありそうだ",
-			"休みながらも好奇心は止まらない",
-			"次の探索ルートを考えよう",
-			"面白い場所だ…もっと調べたい",
-		],
-		floor_reached: [
-			"新しい階層だ！何があるかな",
-			"どんな敵がいるんだろう",
-			"この階層の構造は…",
-			"ワクワクする！探索開始だ",
-			"前の階層とは何が違うかな",
-			"環境の変化を観察しよう",
-			"新しい発見のチャンスだ",
-			"この階層特有の特徴は何だろう",
-			"地形パターンに変化はあるかな",
-			"深くなるほど面白くなる",
-			"ここにはどんな秘密があるんだろう",
-			"新しい生態系があるかもしれない",
-			"探索のしがいがありそうだ",
-			"調査器具の準備は万全だ",
-			"この階層を徹底的に調べたい",
-		],
-		jump_success: [
-			"跳躍！着地点には何が？",
-			"飛んでみたらどうなるかと思って",
-			"この距離を飛べるのか！",
-			"面白い！もっと飛んでみたい",
-			"着地成功！向こう側は…",
-			"跳躍のメカニズムが面白い",
-			"着地点の地形を観察しよう",
-			"飛距離と角度の関係が気になる",
-			"あの位置からここまで飛べるとは",
-			"ジャンプ中の景色が新鮮だった",
-			"この感覚は初めてだ！",
-			"ローブの裾が広がって着地しにくいな",
-			"物理法則が気になる跳躍だった",
-			"次はもっと遠くまで飛べるかな",
-			"着地の衝撃が意外と少ない、なぜだろう",
-		],
-		card_acquired: [
-			"このカード、どんな効果だろう",
-			"面白い！新しい手札だ",
-			"カードの仕組みが気になる",
-			"どう使えば最大限活きるかな",
-			"デッキの組み合わせが広がる！",
-			"このカードの由来が知りたい",
-			"興味深い一枚だ",
-			"手帳にカード情報を記録しよう",
-			"新しい戦術を試せるぞ",
-			"このカードで何ができるだろう",
-			"組み合わせの可能性を考えたい",
-			"デッキ構築って奥が深いな",
-			"ローブのポケットに大事にしまおう",
-			"この一枚が戦略を変えるかも",
-			"研究しがいのあるカードだ",
-		],
-		card_skipped: [
-			"今は見送ろう、理由はある",
-			"このデッキの構成を崩したくない",
-			"今のバランスが興味深いんだ",
-			"次の機会に分析しよう",
-			"見送ったほうが面白い展開になりそう",
-			"今の手札の組み合わせを試したい",
-			"変えないことにも意味がある",
-			"この構成でもっと研究したい",
-			"スキップも戦略の一つだ",
-			"今のデッキに愛着がある",
-			"手帳に記録だけしておこう",
-			"また出会える機会があるだろう",
-			"現状の手札で実験を続けたい",
-			"このデッキの可能性をもっと探りたい",
-			"見送るのも一つの発見だ",
-		],
-	},
-} as const;
-
-/**
- * レア発話データ辞書
- * 各性格の"素"が出るギャップ表現（低確率で出現）
- * デフォルト発話選択時のみ判定される
- */
-export const RARE_SPEECH_VARIANTS: Record<
-	Personality,
-	Partial<Record<SpeechEventType, readonly string[]>>
-> = {
-	brave: {
-		move_success: ["…本当にこの道で合ってるのかな"],
-		enemy_defeated: ["正直、ちょっと怖かった…"],
-		damage_taken: ["痛い…誰か助けて…"],
-		game_over: ["怖かった…ずっと怖かったんだ"],
-		combo_activated: ["手が震えてるけど…いける！"],
-		treasure_found: ["ほっとした…本当に"],
-		trap_triggered: ["もう帰りたい…"],
-		floor_reached: ["この先…大丈夫かな"],
-		card_acquired: ["どれがいいのか…正直わからない"],
-		card_skipped: ["本当にこれでよかったのかな…"],
-	},
-	cautious: {
-		move_success: ["たまには直感で行ってみるか！"],
-		enemy_defeated: ["分析なんていらなかった！"],
-		damage_taken: ["考えてる場合じゃない、走れ！"],
-		game_over: ["計算なんて…無意味だったか"],
-		combo_activated: ["理屈抜きで気持ちいい！"],
-		treasure_found: ["深く考えず、素直に喜ぼう"],
-		trap_triggered: ["えいっ！…って踏んじゃった"],
-		floor_reached: ["今日は勘で進んでみよう"],
-		card_acquired: ["直感で選んじゃおう！"],
-		card_skipped: ["あえて考えずにスキップ！"],
-	},
-	cheerful: {
-		move_success: ["…ここ、少し嫌な予感がする"],
-		enemy_defeated: ["もう誰も傷つかないで"],
-		damage_taken: ["…笑えない、かも"],
-		game_over: ["みんなを守れなかった…ごめん"],
-		combo_activated: ["集中…今は絶対に外せない"],
-		treasure_found: ["これで…もう少し頑張れる"],
-		trap_triggered: ["痛い…けど、泣かないよ"],
-		floor_reached: ["ここからは気を引き締めないと"],
-		card_acquired: ["…このカード、大事にしなきゃ"],
-		card_skipped: ["…本当は欲しかったかも"],
-	},
-	stoic: {
-		move_success: ["…この景色、少し好きだ"],
-		enemy_defeated: ["…よくやった、俺"],
-		damage_taken: ["…正直、きつい"],
-		game_over: ["…もう少し、生きたかった"],
-		combo_activated: ["…悪くない、この感覚"],
-		treasure_found: ["…ありがたいな、本当に"],
-		trap_triggered: ["…くそっ"],
-		floor_reached: ["…少しだけ、わくわくする"],
-		card_acquired: ["…少し、嬉しいかもしれない"],
-		card_skipped: ["…次がある、きっと"],
-	},
-	curious: {
-		move_success: ["この道で間違いない、確信がある"],
-		enemy_defeated: ["分かっていた、こうなると"],
-		damage_taken: ["想定内だ、問題ない"],
-		game_over: ["全ては分かっていたのに…"],
-		combo_activated: ["読み通り、完璧な連携だ"],
-		treasure_found: ["ここにあると思っていた"],
-		trap_triggered: ["知っていた…けど避けられなかった"],
-		floor_reached: ["この階層の構造は把握済みだ"],
-		card_acquired: ["このカードが手に入ると確信していた"],
-		card_skipped: ["最初から見送るつもりだった"],
-	},
-} as const;
-
-/**
- * 連続発話データ辞書
- * キーは "直前イベント_現在イベント" 形式
- * 未定義パターンは通常発話にフォールバック
- */
-export const SPEECH_SEQUENCE_VARIANTS: Record<
-	Personality,
-	Partial<Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>>
-> = {
-	brave: {
-		damage_taken_enemy_defeated: ["やり返したぞ！"],
-		move_fail_move_fail: ["また壁か…剣で斬りたいぜ！"],
-		trap_triggered_move_success: ["今度は慎重に行くぞ"],
-		enemy_defeated_enemy_defeated: ["止まらないぞ！"],
-		attack_miss_enemy_defeated: ["今度は外さなかったな！"],
-		damage_taken_move_success: ["ここは一旦退くぞ"],
-	},
-	cautious: {
-		damage_taken_enemy_defeated: ["反撃成功、被害を取り戻した"],
-		move_fail_move_fail: ["マップの把握が甘いな…眼鏡を拭こう"],
-		trap_triggered_move_success: ["二度と同じ轍は踏まない"],
-		enemy_defeated_enemy_defeated: ["効率よく排除できている"],
-		attack_miss_enemy_defeated: ["修正完了、確実に仕留めた"],
-		damage_taken_move_success: ["被害を最小限に、退避する"],
-	},
-	cheerful: {
-		damage_taken_enemy_defeated: ["やられたらやり返す〜！"],
-		move_fail_move_fail: ["あはは、また壁だ〜！"],
-		trap_triggered_move_success: ["今度は気をつけるぞ〜"],
-		enemy_defeated_enemy_defeated: ["連勝だ〜！すごいすごい！"],
-		attack_miss_enemy_defeated: ["リベンジ成功〜！"],
-		damage_taken_move_success: ["いったん逃げるよ〜！"],
-	},
-	stoic: {
-		damage_taken_enemy_defeated: ["…借りは返した"],
-		move_fail_move_fail: ["…壁ばかりだ"],
-		trap_triggered_move_success: ["…二度はない"],
-		enemy_defeated_enemy_defeated: ["…続けて排除"],
-		attack_miss_enemy_defeated: ["…今度は当てた"],
-		damage_taken_move_success: ["…退く"],
-	},
-	curious: {
-		damage_taken_enemy_defeated: ["さっきの攻撃を分析して倒せた！"],
-		move_fail_move_fail: ["この壁の配置には法則がありそうだ"],
-		trap_triggered_move_success: ["罠の位置を手帳に記録、次は大丈夫"],
-		enemy_defeated_enemy_defeated: ["連戦データが集まるな！"],
-		attack_miss_enemy_defeated: ["なるほど、こう当てればいいのか"],
-		damage_taken_move_success: ["距離を取って観察しよう"],
-	},
-} as const;
