+++ src/game/speechData/brave.ts
@@ -1,0 +1,268 @@
+import type { SpeechEventType } from "../../types";
+
+export const speechVariants: Record<SpeechEventType, readonly string[]> = {
+	move_success: [
+		"よし、進もう",
+		"この先は…",
+		"恐れるな、前へ！",
+		"道は俺が切り拓く",
+		"立ち止まるな！",
+		"迷う必要はない",
+		"俺について来い",
+		"一歩一歩が前進だ",
+		"この先に勝利がある",
+		"足を止めるな！",
+		"さあ、突き進め",
+		"後ろは振り返らない",
+		"行くぞ、仲間たち",
+		"剣を握る手に力が入る",
+		"この体で道を切り拓くぞ",
+	],
+	move_fail: [
+		"行き止まりか…",
+		"ここは通れない",
+		"別の道を探そう",
+		"壁か、ならば回り込む",
+		"引き返すのも勇気だ",
+		"壁一枚で諦めるか！",
+		"くっ、ここは塞がれている",
+		"この壁、剣で斬れないか",
+		"迂回路があるはずだ",
+		"焦るな、道は必ずある",
+		"壁は俺を止められない",
+		"ここじゃない、別の道だ",
+		"力ずくで壊したいところだが",
+		"退くのは負けじゃない",
+		"別の突破口を見つける",
+	],
+	attack_miss: [
+		"空振りだ！",
+		"くっ、外した",
+		"手応えがない…",
+		"次は必ず当てる！",
+		"まだだ、諦めない！",
+		"甘かったか…！",
+		"この程度で終わらん！",
+		"剣筋が甘かった",
+		"もう一撃、叩き込む！",
+		"読まれたか…だが！",
+		"反省は後だ、もう一度！",
+		"焦ったか…冷静にいけ",
+		"敵もやるな…！",
+		"この腕で仕留めてみせる！",
+		"外したが、次は違う！",
+	],
+	combo_activated: [
+		"今だ！畳みかける！",
+		"チャンスだ！",
+		"連続攻撃！",
+		"止まるなよ！",
+		"一気に決める！",
+		"怒涛の連撃だ！",
+		"この勢いで押し切る！",
+		"逃がさないぞ！",
+		"俺の剣が唸る！",
+		"休む暇は与えない！",
+		"全力で叩き込め！",
+		"ここで仕留める！",
+		"連撃で決着だ！",
+		"勝負はここだ！",
+		"剣の舞だ、食らえ！",
+	],
+	enemy_defeated: [
+		"やった！",
+		"倒したぞ",
+		"まだ先がある",
+		"俺の勝ちだ！",
+		"強敵だったな",
+		"これで道が開けた",
+		"次の敵も来い！",
+		"まだまだ行けるぞ",
+		"この剣に斬れぬものはない",
+		"この力、信じてよかった",
+		"一つ片付いたな",
+		"手加減はしない主義だ",
+		"見事な戦いだった",
+		"この体格で押し切ったぞ",
+		"さらに強くなれる",
+	],
+	damage_taken: [
+		"くっ…！",
+		"痛い！",
+		"まだ耐えられる",
+		"この程度で倒れるか！",
+		"効いたが…まだだ！",
+		"かすり傷だ！",
+		"俺はまだ立てる！",
+		"この体はそう簡単には折れん！",
+		"歯を食いしばれ…！",
+		"ここで倒れてたまるか",
+		"傷が増えるほど強くなる！",
+		"痛みは戦士の勲章だ",
+		"やるな…だが！",
+		"膝はまだ折れていない！",
+		"根性で耐えてみせる！",
+	],
+	game_over: [
+		"ここまでか…",
+		"無念…",
+		"次こそは…",
+		"悔しい…だが良い戦いだった",
+		"俺はまだ…終わらない…",
+		"力が…足りなかった",
+		"すまない…仲間たち",
+		"この敗北を糧にする",
+		"まだ倒れるわけには…",
+		"もっと強くならねば",
+		"剣を握る力がもう…",
+		"次は…必ず勝つ",
+		"この悔しさ、忘れない",
+		"立ち上がれ…俺",
+		"この体が動く限り、戦いは終わらない…",
+	],
+	trap_triggered: [
+		"しまった、罠だ！",
+		"うわっ！",
+		"不覚…！",
+		"くそ、引っかかった！",
+		"罠ごときに…！",
+		"油断した…！",
+		"こんなもので俺を！",
+		"足元を見てなかった…！",
+		"罠師め…やるな",
+		"次は踏まんぞ！",
+		"この屈辱は忘れん！",
+		"卑怯な手を…！",
+		"ぐっ…この体格じゃ避けにくいな",
+		"剣で斬れる罠ならいいのに",
+		"二度目はないぞ！",
+	],
+	treasure_found: [
+		"お、宝箱だ！",
+		"ありがたい",
+		"回復できた",
+		"よし、これで態勢を立て直せる",
+		"運も実力のうちだ！",
+		"天が俺に味方した！",
+		"これで戦える！",
+		"戦士への褒美だな",
+		"助かる、使わせてもらう",
+		"この先の戦いに備えよう",
+		"よし、まだやれるぞ",
+		"思わぬ収穫だ",
+		"剣を振るう力が戻った",
+		"これがあれば心強い",
+		"戦いの神に感謝だ",
+	],
+	rest_area_used: [
+		"少し休もう",
+		"体力が戻った",
+		"ふぅ、一息つける",
+		"休息も戦いの一部だ",
+		"英気を養うぞ",
+		"ここで力を蓄える",
+		"体が軽くなった",
+		"束の間の休息だな",
+		"剣の手入れもしておくか",
+		"次の戦いに備える",
+		"安らげる場所があるとは",
+		"この体を万全にしておくぞ",
+		"万全の状態で挑む",
+		"戦士にも休息は必要だ",
+		"よし、準備完了だ",
+	],
+	floor_reached: [
+		"新しい階層だ",
+		"ここからが本番",
+		"気を引き締めよう",
+		"さあ、行くぞ！",
+		"どんな敵が待っていようと！",
+		"新たな戦場だ！",
+		"さらなる試練か、望むところだ",
+		"この先には何がある",
+		"俺の剣が試される時だ",
+		"強敵よ、かかって来い！",
+		"階層が変わっても俺は変わらん",
+		"深く潜るほど燃えてくる",
+		"ここで引き返す男じゃない",
+		"新たな冒険の始まりだ",
+		"覚悟はできている",
+	],
+	jump_success: [
+		"突っ込むぞ！",
+		"一気に行く！",
+		"ここを抜ける！",
+		"飛び込め！",
+		"勢いで突破だ！",
+		"飛んだ！着地！",
+		"大胆に行くぞ！",
+		"この体重で跳べるとはな！",
+		"勢いは止まらない！",
+		"華麗に決めた！",
+		"飛び越えてやったぞ！",
+		"この程度、朝飯前だ",
+		"勇気ある一歩だ！",
+		"躊躇なく飛ぶ！",
+		"行ける、行けるぞ！",
+	],
+	card_acquired: [
+		"新しい力だ！",
+		"これは使えるぞ！",
+		"戦力強化だ！",
+		"この手札で戦い抜く！",
+		"頼もしい一枚だ",
+		"よし、手に入れたぞ！",
+		"これで戦い方が広がる！",
+		"強くなれる気がする！",
+		"俺の手札に加わるか！",
+		"この力、受け取った！",
+		"新しい武器だな！",
+		"デッキが強化された！",
+		"いい獲物だ！",
+		"この一枚で勝ちに行く！",
+		"戦いの幅が広がるぞ！",
+	],
+	card_skipped: [
+		"今の手札で十分だ",
+		"必要ないな",
+		"俺にはこの手札がある",
+		"見送るぞ",
+		"今のデッキを信じる",
+		"手札の完成度は十分だ",
+		"これ以上は不要だ",
+		"今の構成で戦い抜く",
+		"欲張りすぎは禁物だ",
+		"この手札が俺の武器だ",
+		"変える必要はない",
+		"温存しておく",
+		"今は見送りだ",
+		"現状維持で十分強い",
+		"このデッキで突き進む！",
+	],
+};
+
+export const rareSpeechVariants: Partial<
+	Record<SpeechEventType, readonly string[]>
+> = {
+	move_success: ["…本当にこの道で合ってるのかな"],
+	enemy_defeated: ["正直、ちょっと怖かった…"],
+	damage_taken: ["痛い…誰か助けて…"],
+	game_over: ["怖かった…ずっと怖かったんだ"],
+	combo_activated: ["手が震えてるけど…いける！"],
+	treasure_found: ["ほっとした…本当に"],
+	trap_triggered: ["もう帰りたい…"],
+	floor_reached: ["この先…大丈夫かな"],
+	card_acquired: ["どれがいいのか…正直わからない"],
+	card_skipped: ["本当にこれでよかったのかな…"],
+};
+
+export const sequenceVariants: Partial<
+	Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>
+> = {
+	damage_taken_enemy_defeated: ["やり返したぞ！"],
+	move_fail_move_fail: ["また壁か…剣で斬りたいぜ！"],
+	trap_triggered_move_success: ["今度は慎重に行くぞ"],
+	enemy_defeated_enemy_defeated: ["止まらないぞ！"],
+	attack_miss_enemy_defeated: ["今度は外さなかったな！"],
+	damage_taken_move_success: ["ここは一旦退くぞ"],
+};
