+++ src/game/speechData/cautious.ts
@@ -1,0 +1,268 @@
+import type { SpeechEventType } from "../../types";
+
+export const speechVariants: Record<SpeechEventType, readonly string[]> = {
+	move_success: [
+		"周囲を確認…問題ない",
+		"一歩ずつ、慎重に",
+		"罠がないか注意しよう",
+		"安全を確認して進む",
+		"この方向で合っているはず",
+		"地形に変化はないな",
+		"視界良好、進行可能",
+		"足元を確認…安全だ",
+		"次の一手を考えながら進む",
+		"予定ルート通りだ",
+		"周辺の敵影なし、前進する",
+		"リスクは最小限に抑えた",
+		"眼鏡を直しつつ…地形を確認",
+		"想定内の道だ",
+		"慎重に、だが確実に",
+	],
+	move_fail: [
+		"予想通り、通れないな",
+		"別ルートを検討しよう",
+		"ここは危険かもしれない",
+		"想定内だ、迂回する",
+		"事前に確認しておくべきだった",
+		"代替ルートに切り替える",
+		"壁の構造を記憶しておこう",
+		"想定外だが対処可能だ",
+		"ここは通行不可、記録した",
+		"マップの修正が必要だ",
+		"予備プランに移行する",
+		"眼鏡越しに壁を観察…素材は石か",
+		"情報が足りなかったか",
+		"安全な迂回路を探す",
+		"壁の位置を覚えておく",
+	],
+	attack_miss: [
+		"読みが外れた…",
+		"距離を見誤ったか",
+		"次は確実に仕留める",
+		"焦りは禁物だ",
+		"冷静に、もう一度",
+		"敵の動きを再分析する",
+		"角度の計算を修正しよう",
+		"パターンを読み直す",
+		"誤差の原因を特定する",
+		"想定と実際のずれか",
+		"眼鏡がずれた…修正して再試行",
+		"この失敗から学べることがある",
+		"感情に流されるな",
+		"集中力を維持しろ",
+		"次こそ精度を上げる",
+	],
+	combo_activated: [
+		"計算通りだ",
+		"読み通りの展開",
+		"ここが好機、畳みかける",
+		"想定した連携だ",
+		"隙を見逃さない",
+		"シミュレーション通りの結果だ",
+		"予測していた展開だ",
+		"ここで押し切るのが最善",
+		"効率的に攻撃を重ねる",
+		"全て計画の範囲内だ",
+		"リスクを最小化しつつ攻める",
+		"分析が功を奏した",
+		"最適解で攻め続ける",
+		"冷静に連撃を叩き込む",
+		"眼鏡の奥で勝機を捉えた",
+	],
+	enemy_defeated: [
+		"予定通り排除した",
+		"脅威を一つ除去",
+		"油断はまだ早い",
+		"次の敵に備えよう",
+		"分析通りの弱点だった",
+		"被害を最小限に抑えられた",
+		"想定時間内に排除完了",
+		"このパターンは記憶した",
+		"次はもっと効率よく倒せる",
+		"周囲に残存敵がいないか確認",
+		"データを蓄積しておこう",
+		"勝因を分析しておく",
+		"予備の戦術は不要だったな",
+		"計画通りの戦闘結果だ",
+		"冷静な判断が勝利を呼んだ",
+	],
+	damage_taken: [
+		"想定よりダメージが大きい",
+		"防御が甘かった…",
+		"回復手段を確認しないと",
+		"ここは退くべきか…",
+		"被害を最小限に抑えねば",
+		"被弾原因を分析する",
+		"体力の残量を確認…",
+		"この攻撃パターンを記憶した",
+		"防御態勢を見直す必要がある",
+		"リスク評価を修正しなければ",
+		"眼鏡が…割れてなければいいが",
+		"ダメージ量は許容範囲内か…",
+		"痛手だが致命的ではない",
+		"立て直しの余地はまだある",
+		"撤退ラインを再設定する",
+	],
+	game_over: [
+		"判断を誤ったか…",
+		"もっと慎重にすべきだった",
+		"データは次に活かせる",
+		"敗因を分析しなければ…",
+		"この経験は無駄にしない",
+		"リスク評価が甘かった…",
+		"見落としがあったか…",
+		"より精密な計画が必要だった",
+		"仮説が間違っていた…",
+		"データ不足が敗因だ",
+		"撤退タイミングを逸した…",
+		"冷静さを欠いたか…",
+		"次回はこの失敗を糧にする",
+		"分析し直さなければ…",
+		"全てのパターンを想定できなかった",
+	],
+	trap_triggered: [
+		"見落としていた…不覚",
+		"罠の兆候を見逃したか",
+		"次からはもっと注意を",
+		"確認が甘かった…",
+		"この種の罠は記憶した",
+		"チェック手順に漏れがあった",
+		"この罠のパターンを記録する",
+		"注意力が散漫だったか…",
+		"安全確認を怠った結果だ",
+		"同種の罠を警戒リストに追加",
+		"眼鏡を拭いていたら踏んだ…",
+		"もう一段階慎重にすべきだった",
+		"この失敗は二度と繰り返さない",
+		"予想外の罠の設置位置だ",
+		"探知方法を改善しなければ",
+	],
+	treasure_found: [
+		"回復アイテムか、助かる",
+		"今の状態なら使い時だ",
+		"貴重な回復の機会だ",
+		"ありがたい、備蓄しておきたいが",
+		"これで少し余裕ができた",
+		"状況改善の好機だ",
+		"使用タイミングを見極めよう",
+		"予備のリソースが増えた",
+		"これは戦略的に有利になる",
+		"残り体力と相談して使おう",
+		"計画の修正が必要かもしれない",
+		"ここに宝箱がある確率は考慮済み",
+		"有効活用できるよう計画しよう",
+		"危機的状況の回避に役立つ",
+		"回復分を計算に組み込もう",
+	],
+	rest_area_used: [
+		"ここで態勢を整えよう",
+		"安全を確認してから休む",
+		"回復しつつ次の計画を練る",
+		"体力管理は最重要だ",
+		"焦らず回復を優先する",
+		"休息の間にデータを整理する",
+		"次の行動計画を立てよう",
+		"ここが安全地帯なら最大限活用する",
+		"体力と精神力を回復させる",
+		"先のルートを検討しておこう",
+		"冷静になれる貴重な時間だ",
+		"眼鏡を磨きつつ情報を整理する",
+		"体調管理も戦略の一部だ",
+		"しっかり休んで万全を期す",
+		"回復後の行動方針を固めておく",
+	],
+	floor_reached: [
+		"新しい階層か、偵察から始めよう",
+		"まず地形を把握する",
+		"未知の敵に警戒を",
+		"情報収集が先だ",
+		"慎重に進めていこう",
+		"新たなデータ収集の機会だ",
+		"前の階層の知見を活かそう",
+		"敵の配置パターンを予測する",
+		"安全な拠点を確保したい",
+		"環境の変化に適応しなければ",
+		"警戒レベルを上げておく",
+		"先入観を排除して観察しよう",
+		"眼鏡を直して…新たな分析開始",
+		"リスク要因を洗い出そう",
+		"あらゆる可能性に備えておく",
+	],
+	jump_success: [
+		"計算通り、着地成功",
+		"リスクはあったが通過できた",
+		"飛距離は想定内だ",
+		"着地点の安全を確認",
+		"大胆な手だが必要だった",
+		"着地の衝撃は許容範囲内",
+		"予測座標に正確に着地",
+		"この手段が最善だった",
+		"想定通りの軌道だ",
+		"安全マージンは確保できた",
+		"着地後の周囲を確認する",
+		"計算を信じて飛んだ甲斐があった",
+		"リスクとリターンの判断は正しかった",
+		"次も同じ判断ができるように記録する",
+		"飛距離のデータを記録しておこう",
+	],
+	card_acquired: [
+		"カードの性能を確認しよう",
+		"デッキの最適化が可能だ",
+		"戦力の見直しが必要だ",
+		"この一枚の価値を分析する",
+		"手札構成の改善に使えるか",
+		"有効な選択肢が増えた",
+		"デッキバランスを考慮しよう",
+		"期待値が上がる組み合わせだ",
+		"この追加は計画通りだ",
+		"戦術の幅が広がる",
+		"慎重に入れ替えを検討する",
+		"リスクとリターンを天秤にかけよう",
+		"シナジーのある構成を目指す",
+		"この一枚でデッキが完成に近づく",
+		"眼鏡越しにカードを精査する",
+	],
+	card_skipped: [
+		"現状のデッキが最適解だ",
+		"入れ替えるメリットが薄い",
+		"無理に変更する必要はない",
+		"リスクを増やすべきではない",
+		"今の構成が安定している",
+		"不要な変更は避ける",
+		"現状維持が最善の判断だ",
+		"デッキの安定性を優先する",
+		"変更による損失リスクが高い",
+		"見送りが合理的な選択だ",
+		"この判断は間違っていない",
+		"分析の結果、不要と判断した",
+		"今のバランスを崩したくない",
+		"次の機会を待とう",
+		"冷静な判断ができた",
+	],
+};
+
+export const rareSpeechVariants: Partial<
+	Record<SpeechEventType, readonly string[]>
+> = {
+	move_success: ["たまには直感で行ってみるか！"],
+	enemy_defeated: ["分析なんていらなかった！"],
+	damage_taken: ["考えてる場合じゃない、走れ！"],
+	game_over: ["計算なんて…無意味だったか"],
+	combo_activated: ["理屈抜きで気持ちいい！"],
+	treasure_found: ["深く考えず、素直に喜ぼう"],
+	trap_triggered: ["えいっ！…って踏んじゃった"],
+	floor_reached: ["今日は勘で進んでみよう"],
+	card_acquired: ["直感で選んじゃおう！"],
+	card_skipped: ["あえて考えずにスキップ！"],
+};
+
+export const sequenceVariants: Partial<
+	Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>
+> = {
+	damage_taken_enemy_defeated: ["反撃成功、被害を取り戻した"],
+	move_fail_move_fail: ["マップの把握が甘いな…眼鏡を拭こう"],
+	trap_triggered_move_success: ["二度と同じ轍は踏まない"],
+	enemy_defeated_enemy_defeated: ["効率よく排除できている"],
+	attack_miss_enemy_defeated: ["修正完了、確実に仕留めた"],
+	damage_taken_move_success: ["被害を最小限に、退避する"],
+};
