+++ src/game/speechData/cheerful.ts
@@ -1,0 +1,268 @@
+import type { SpeechEventType } from "../../types";
+
+export const speechVariants: Record<SpeechEventType, readonly string[]> = {
+	move_success: [
+		"さくさく進むよ〜！",
+		"いい感じ！",
+		"次はどこかな〜",
+		"わくわくするね！",
+		"どんどん行こう！",
+		"ルンルン♪",
+		"お散歩気分〜！",
+		"今日はいい日だな〜",
+		"この道、好きかも！",
+		"快調快調〜！",
+		"冒険って楽しい！",
+		"小さい体でもどんどん進むよ！",
+		"足取り軽いよ〜",
+		"狭い隙間もスイスイだね！",
+		"たのしいな〜！",
+	],
+	move_fail: [
+		"あれ〜、行き止まり！",
+		"んー、こっちじゃないか",
+		"まあ、別の道もあるし！",
+		"ドンマイドンマイ！",
+		"あはは、戻ろう！",
+		"おっとっと、壁だ〜",
+		"行き止まりも冒険のうち！",
+		"残念〜、でも大丈夫！",
+		"違う道にも何かあるかも！",
+		"ここは壁だね〜",
+		"まあまあ、気にしない！",
+		"回り道も楽しいよ！",
+		"はずれだ〜、次いこ次！",
+		"壁さんこんにちは〜！",
+		"小さいから隙間ないか探してみよ〜",
+	],
+	attack_miss: [
+		"あちゃー、外しちゃった！",
+		"惜しい〜！",
+		"次は当てるぞ〜！",
+		"えへへ、ごめんごめん",
+		"もう一回！",
+		"あらら〜、空振り！",
+		"ドンマイ自分！",
+		"惜しかったのに〜！",
+		"次こそは！えいっ！",
+		"スカッと空振り〜！",
+		"あたらなかった〜",
+		"リーチが足りないかな〜！",
+		"えーい、待ってよ〜！",
+		"がんばれ自分！",
+		"外しちゃったけど元気！",
+	],
+	combo_activated: [
+		"やったやった！連続だ！",
+		"ノリノリだね！",
+		"すごいすごい！",
+		"いっけー！",
+		"コンボだ〜！楽しい！",
+		"テンション上がる〜！",
+		"絶好調だよ〜！",
+		"どんどんいくよ〜！",
+		"連続ヒット〜！最高！",
+		"止まらないよ〜！",
+		"コンボって気持ちいい！",
+		"きたきたきた〜！",
+		"小さくても強いんだから〜！",
+		"もっともっと！",
+		"この感じ、サイコー！",
+	],
+	enemy_defeated: [
+		"やったね！",
+		"勝った勝った〜！",
+		"ナイスファイト！",
+		"いえーい！",
+		"よっしゃー！",
+		"お疲れさまでした〜！",
+		"勝っちゃった〜！",
+		"ばんざーい！",
+		"楽しかった〜！",
+		"すっきり〜！",
+		"また勝っちゃった！",
+		"小さい体でもやるでしょ！",
+		"大勝利〜！",
+		"ナイスナイス！",
+		"これで安心だね！",
+	],
+	damage_taken: [
+		"いたた！",
+		"うぅ、やるね〜",
+		"まだまだ元気だよ！",
+		"ちょっと痛かった…えへへ",
+		"大丈夫大丈夫！",
+		"あいたた〜！",
+		"やられた〜、でも平気！",
+		"ちくっとしたかな？",
+		"負けないもんね〜！",
+		"痛いけどへっちゃら！",
+		"ぐぬぬ…でも笑顔！",
+		"小さい体にはちょっと響くかも〜",
+		"これくらいなんともない！",
+		"ちょっとびっくりしただけ！",
+		"痛いの痛いの飛んでけ〜！",
+	],
+	game_over: [
+		"あちゃ〜、やられちゃった",
+		"また挑戦しよう！",
+		"次はきっとうまくいくよ！",
+		"楽しかった〜！もう一回！",
+		"ドンマイ！切り替えていこう",
+		"おしかったな〜",
+		"次こそ頑張る〜！",
+		"いい勉強になったね！",
+		"まだまだこれから！",
+		"失敗は成功のもとだよ！",
+		"今度はもっと上手にやるぞ〜",
+		"やられちゃったけど楽しかった！",
+		"ま、こういう日もあるよね！",
+		"またすぐ戻ってくるよ〜！",
+		"リベンジだ〜！",
+	],
+	trap_triggered: [
+		"きゃっ！罠だった！",
+		"びっくりした〜！",
+		"わわっ、やられた！",
+		"あはは、引っかかっちゃった",
+		"もう〜、意地悪な罠！",
+		"うわわ〜！罠だ〜！",
+		"ひゃっ！びっくり！",
+		"やだやだ〜、罠だなんて！",
+		"あちゃ〜、踏んじゃった",
+		"いたっ！もう〜！",
+		"サプライズは嫌いだよ〜",
+		"小さいから見えなかったよ〜",
+		"次は気をつけるね〜！",
+		"えーん、罠にやられた〜",
+		"ドッキリかと思った〜！",
+	],
+	treasure_found: [
+		"わーい、宝箱だ！",
+		"ラッキー！",
+		"やった〜、うれしい！",
+		"ご褒美だね！",
+		"いいもの見つけた〜！",
+		"宝箱だ〜！開けちゃお！",
+		"わくわく、何が入ってるかな！",
+		"きらきら〜！",
+		"プレゼントみたい！",
+		"うれしいうれしい！",
+		"今日はついてるね〜！",
+		"ハッピー！",
+		"これはうれしいサプライズ！",
+		"お宝ゲット〜！",
+		"大当たり〜！",
+	],
+	rest_area_used: [
+		"ちょっと休憩〜！",
+		"はぁ〜、気持ちいい！",
+		"元気回復！",
+		"リフレッシュ完了！",
+		"よーし、また頑張ろ！",
+		"のんびりタイム〜！",
+		"ほっと一息〜！",
+		"疲れが吹き飛んだ〜！",
+		"ごろんってしたいな〜",
+		"ぽかぽかする〜",
+		"充電完了！",
+		"小さい体はすぐ回復するよ〜！",
+		"休憩って最高！",
+		"体が軽くなった〜！",
+		"エネルギー満タン！",
+	],
+	floor_reached: [
+		"新しいフロアだ〜！",
+		"わくわく！何があるかな",
+		"楽しくなってきた！",
+		"どんどん奥へ！",
+		"冒険の続きだね！",
+		"新しい景色だ〜！",
+		"探検探検〜！",
+		"ここはどんなところかな！",
+		"テンション上がる〜！",
+		"もっと深くまで行くよ！",
+		"新ステージだ〜！",
+		"きゃー、楽しみ！",
+		"ドキドキワクワク！",
+		"新しい出会いがあるかな〜",
+		"わーい、次のフロア！",
+	],
+	jump_success: [
+		"とりゃー！",
+		"ジャンプ成功！",
+		"飛んだ飛んだ〜！",
+		"えいっ！着地！",
+		"気持ちいい〜！",
+		"ぴょーん！",
+		"着地成功〜！",
+		"飛ぶの楽しい〜！",
+		"身軽でしょ〜！",
+		"ひゅーん！とんだ〜！",
+		"やった〜、飛べた！",
+		"ジャンプ大好き！",
+		"小さい体は軽いから楽チン！",
+		"すごい跳んだ〜！",
+		"ばっちり着地！",
+	],
+	card_acquired: [
+		"わ〜い、新しいカード！",
+		"どれと交換しようかな〜！",
+		"きらきらしてる〜！",
+		"いいの見つけた〜！",
+		"カードゲットだよ〜！",
+		"やった〜、新戦力！",
+		"これ強そう〜！",
+		"うれしいうれしい！",
+		"デッキが賑やかになるね！",
+		"どれにしよっかな〜！",
+		"わくわくする〜！",
+		"お宝カードだ〜！",
+		"ラッキー！いいカード！",
+		"これは使えそう〜！",
+		"新しいカードって楽しいね！",
+	],
+	card_skipped: [
+		"んー、今はいいかな〜！",
+		"またの機会にね〜！",
+		"今のデッキが好きだもん！",
+		"パスパス〜！",
+		"今回はいいや〜！",
+		"今の手札がお気に入り〜！",
+		"見送っちゃおう！",
+		"ドンマイ〜、次があるよ！",
+		"今は変えなくて大丈夫！",
+		"このままでいっこう〜！",
+		"スキップだよ〜！",
+		"今の手札で頑張る〜！",
+		"また次のチャンスがあるし！",
+		"今はこのままでオッケー！",
+		"お気に入りは変えないもん〜！",
+	],
+};
+
+export const rareSpeechVariants: Partial<
+	Record<SpeechEventType, readonly string[]>
+> = {
+	move_success: ["…ここ、少し嫌な予感がする"],
+	enemy_defeated: ["もう誰も傷つかないで"],
+	damage_taken: ["…笑えない、かも"],
+	game_over: ["みんなを守れなかった…ごめん"],
+	combo_activated: ["集中…今は絶対に外せない"],
+	treasure_found: ["これで…もう少し頑張れる"],
+	trap_triggered: ["痛い…けど、泣かないよ"],
+	floor_reached: ["ここからは気を引き締めないと"],
+	card_acquired: ["…このカード、大事にしなきゃ"],
+	card_skipped: ["…本当は欲しかったかも"],
+};
+
+export const sequenceVariants: Partial<
+	Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>
+> = {
+	damage_taken_enemy_defeated: ["やられたらやり返す〜！"],
+	move_fail_move_fail: ["あはは、また壁だ〜！"],
+	trap_triggered_move_success: ["今度は気をつけるぞ〜"],
+	enemy_defeated_enemy_defeated: ["連勝だ〜！すごいすごい！"],
+	attack_miss_enemy_defeated: ["リベンジ成功〜！"],
+	damage_taken_move_success: ["いったん逃げるよ〜！"],
+};
