+++ src/game/speechData/curious.ts
@@ -1,0 +1,268 @@
+import type { SpeechEventType } from "../../types";
+
+export const speechVariants: Record<SpeechEventType, readonly string[]> = {
+	move_success: [
+		"この先に何があるだろう",
+		"面白い構造だな",
+		"どんな発見があるかな",
+		"もっと奥へ行ってみよう",
+		"この通路は気になるな",
+		"壁の模様が独特だ",
+		"この道の先が気になる",
+		"新しい発見の予感がする",
+		"地形のパターンが見えてきた",
+		"ここの空気が変わった気がする",
+		"なぜこの通路はこの形なんだろう",
+		"手帳に地形を書き留めておこう",
+		"この建築様式は興味深い",
+		"まだまだ知らないことだらけだ",
+		"観察しながら進もう",
+	],
+	move_fail: [
+		"ここは行けないのか…残念",
+		"壁の向こうには何が？",
+		"別の道に何かあるかも",
+		"なぜここは塞がれてるんだろう",
+		"迂回路を探してみよう",
+		"この壁の材質が気になる",
+		"行き止まりにも意味があるはず",
+		"壁の厚さはどれくらいだろう",
+		"なぜここだけ通れないのか",
+		"壁の向こうの構造が知りたい",
+		"行き止まりか、標本採取でもしておこう",
+		"この配置にはパターンがあるかも",
+		"壁にも何かヒントがあるかもしれない",
+		"塞がれている理由が気になる",
+		"ふむ、ここは通れないのか",
+	],
+	attack_miss: [
+		"なぜ外れた？距離かな",
+		"この敵の回避パターンは…",
+		"どうすれば当たるだろう",
+		"興味深い動きをするな",
+		"次はタイミングを変えてみよう",
+		"この敵の反射速度は予想以上だ",
+		"攻撃の軌道を分析してみよう",
+		"なるほど、そう避けるのか",
+		"回避のメカニズムが気になる",
+		"面白い、別の角度から試そう",
+		"どの距離なら当たるだろう",
+		"この敵の回避能力は論文にできそうだ",
+		"外れた原因を突き止めたい",
+		"攻撃方法を変えてみようか",
+		"データが一つ増えた",
+	],
+	combo_activated: [
+		"この組み合わせが効くのか！",
+		"面白い！連鎖が起きた！",
+		"なるほど、こう繋がるのか",
+		"予想以上の効果だ！",
+		"もっと試してみたいな",
+		"この連携パターンは記録しておこう",
+		"コンボのメカニズムが見えた！",
+		"どこまで連鎖するか試したい",
+		"興味深い相乗効果だ",
+		"この発見は大きいぞ",
+		"もっと効率的な組み合わせがあるかも",
+		"連鎖の法則が分かってきた",
+		"素晴らしい、仮説が実証された",
+		"このパターン、応用できそうだ",
+		"手帳に書ききれないくらいの発見だ",
+	],
+	enemy_defeated: [
+		"この敵は何者だったんだろう",
+		"倒せた！弱点はあそこか",
+		"興味深い戦闘だった",
+		"もっと強い敵もいるのかな",
+		"どんな仕組みで動いてたんだろう",
+		"この敵の生態が気になる",
+		"戦闘データが蓄積されていく",
+		"次はもっと効率的に倒せるかも",
+		"この敵固有の特徴は何だったんだろう",
+		"貴重な観察の機会だった",
+		"倒してしまったか…もう少し観察したかった",
+		"戦闘パターンの分析が進んだ",
+		"似た敵と比較してみたいな",
+		"この敵から学べることは多い",
+		"調査ノートに詳しく記録しておこう",
+	],
+	damage_taken: [
+		"この攻撃はどういう原理だ？",
+		"痛いけど…学びがある",
+		"なるほど、そう来るか",
+		"この敵の攻撃パターンは…",
+		"次は避けられるかな",
+		"この痛みの原因を分析しよう",
+		"攻撃の威力と速度を計測したい",
+		"面白い攻撃方法だな",
+		"被弾データとして記録しておこう",
+		"なぜこのタイミングで攻撃が来た？",
+		"この攻撃の仕組みが気になる",
+		"痛いが、興味深いデータだ",
+		"防御の改善点が見えてきた",
+		"ローブが破れた…繕わないと",
+		"次は対策を立てられそうだ",
+	],
+	game_over: [
+		"まだ知りたいことがあったのに…",
+		"この先に何があったんだろう",
+		"もう少し調べたかった…",
+		"敗因は何だったんだろう",
+		"次はもっとうまくやれるはず",
+		"途中で終わるなんて残念だ",
+		"集めたデータは無駄にしない",
+		"まだ解明できていないことがある",
+		"この経験から何を学べるか…",
+		"もっと観察する時間が欲しかった",
+		"答えに辿り着けなかった…",
+		"次こそ最奥部まで行きたい",
+		"未知の領域がまだたくさんある",
+		"失敗の原因を究明しなければ",
+		"好奇心は尽きない…また挑もう",
+	],
+	trap_triggered: [
+		"これは…罠か！どんな仕組みだ？",
+		"巧妙な罠だ、誰が作ったんだろう",
+		"この罠の構造が気になる",
+		"なるほど、ここに仕掛けてあったか",
+		"次は見抜けるかもしれない",
+		"この罠の起動メカニズムは…",
+		"痛いけど、仕掛けは見事だ",
+		"どんな技術で作られたんだろう",
+		"罠の設計思想が興味深い",
+		"トリガーの位置を記憶した",
+		"同じ原理の罠が他にもあるかも",
+		"この罠、分解して調べたいな",
+		"予想外の場所に仕掛けてあった",
+		"ローブの裾が引っかかったみたい",
+		"これは新しい発見だ…痛いけど",
+	],
+	treasure_found: [
+		"おっ、何が入ってるかな",
+		"これは珍しい！",
+		"どこから来た宝なんだろう",
+		"中身が気になる！",
+		"こんなところに宝箱が…不思議だ",
+		"誰がここに置いたんだろう",
+		"宝箱の装飾が興味深い",
+		"中身を詳しく調べたいな",
+		"この宝箱の由来が気になる",
+		"なぜこの場所に宝があるのか",
+		"貴重な発見だ！",
+		"宝箱のロック機構が面白い",
+		"この素材はどこ産だろう",
+		"回復できる…仕組みが知りたい",
+		"宝箱にもいろいろ種類があるんだな",
+	],
+	rest_area_used: [
+		"ここはなぜ安全なんだろう",
+		"休みながら考えよう",
+		"この場所の成り立ちが気になる",
+		"回復しつつ次の探索を計画",
+		"不思議な空間だな…",
+		"なぜここだけ敵がいないんだろう",
+		"この場所の構造を観察しておこう",
+		"休息中に手帳を整理しよう",
+		"安全地帯の原理が気になる",
+		"ここの空気が他と違う気がする",
+		"回復の仕組みは何なんだろう",
+		"この場所にも秘密がありそうだ",
+		"休みながらも好奇心は止まらない",
+		"次の探索ルートを考えよう",
+		"面白い場所だ…もっと調べたい",
+	],
+	floor_reached: [
+		"新しい階層だ！何があるかな",
+		"どんな敵がいるんだろう",
+		"この階層の構造は…",
+		"ワクワクする！探索開始だ",
+		"前の階層とは何が違うかな",
+		"環境の変化を観察しよう",
+		"新しい発見のチャンスだ",
+		"この階層特有の特徴は何だろう",
+		"地形パターンに変化はあるかな",
+		"深くなるほど面白くなる",
+		"ここにはどんな秘密があるんだろう",
+		"新しい生態系があるかもしれない",
+		"探索のしがいがありそうだ",
+		"調査器具の準備は万全だ",
+		"この階層を徹底的に調べたい",
+	],
+	jump_success: [
+		"跳躍！着地点には何が？",
+		"飛んでみたらどうなるかと思って",
+		"この距離を飛べるのか！",
+		"面白い！もっと飛んでみたい",
+		"着地成功！向こう側は…",
+		"跳躍のメカニズムが面白い",
+		"着地点の地形を観察しよう",
+		"飛距離と角度の関係が気になる",
+		"あの位置からここまで飛べるとは",
+		"ジャンプ中の景色が新鮮だった",
+		"この感覚は初めてだ！",
+		"ローブの裾が広がって着地しにくいな",
+		"物理法則が気になる跳躍だった",
+		"次はもっと遠くまで飛べるかな",
+		"着地の衝撃が意外と少ない、なぜだろう",
+	],
+	card_acquired: [
+		"このカード、どんな効果だろう",
+		"面白い！新しい手札だ",
+		"カードの仕組みが気になる",
+		"どう使えば最大限活きるかな",
+		"デッキの組み合わせが広がる！",
+		"このカードの由来が知りたい",
+		"興味深い一枚だ",
+		"手帳にカード情報を記録しよう",
+		"新しい戦術を試せるぞ",
+		"このカードで何ができるだろう",
+		"組み合わせの可能性を考えたい",
+		"デッキ構築って奥が深いな",
+		"ローブのポケットに大事にしまおう",
+		"この一枚が戦略を変えるかも",
+		"研究しがいのあるカードだ",
+	],
+	card_skipped: [
+		"今は見送ろう、理由はある",
+		"このデッキの構成を崩したくない",
+		"今のバランスが興味深いんだ",
+		"次の機会に分析しよう",
+		"見送ったほうが面白い展開になりそう",
+		"今の手札の組み合わせを試したい",
+		"変えないことにも意味がある",
+		"この構成でもっと研究したい",
+		"スキップも戦略の一つだ",
+		"今のデッキに愛着がある",
+		"手帳に記録だけしておこう",
+		"また出会える機会があるだろう",
+		"現状の手札で実験を続けたい",
+		"このデッキの可能性をもっと探りたい",
+		"見送るのも一つの発見だ",
+	],
+};
+
+export const rareSpeechVariants: Partial<
+	Record<SpeechEventType, readonly string[]>
+> = {
+	move_success: ["この道で間違いない、確信がある"],
+	enemy_defeated: ["分かっていた、こうなると"],
+	damage_taken: ["想定内だ、問題ない"],
+	game_over: ["全ては分かっていたのに…"],
+	combo_activated: ["読み通り、完璧な連携だ"],
+	treasure_found: ["ここにあると思っていた"],
+	trap_triggered: ["知っていた…けど避けられなかった"],
+	floor_reached: ["この階層の構造は把握済みだ"],
+	card_acquired: ["このカードが手に入ると確信していた"],
+	card_skipped: ["最初から見送るつもりだった"],
+};
+
+export const sequenceVariants: Partial<
+	Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>
+> = {
+	damage_taken_enemy_defeated: ["さっきの攻撃を分析して倒せた！"],
+	move_fail_move_fail: ["この壁の配置には法則がありそうだ"],
+	trap_triggered_move_success: ["罠の位置を手帳に記録、次は大丈夫"],
+	enemy_defeated_enemy_defeated: ["連戦データが集まるな！"],
+	attack_miss_enemy_defeated: ["なるほど、こう当てればいいのか"],
+	damage_taken_move_success: ["距離を取って観察しよう"],
+};
