+++ src/game/speechData/stoic.ts
@@ -1,0 +1,268 @@
+import type { SpeechEventType } from "../../types";
+
+export const speechVariants: Record<SpeechEventType, readonly string[]> = {
+	move_success: [
+		"…前進",
+		"進む",
+		"問題ない",
+		"…",
+		"道は続く",
+		"…行ける",
+		"静かだ",
+		"…先へ",
+		"足音だけが響く",
+		"…進行",
+		"道がある",
+		"…歩く",
+		"前へ",
+		"…昔もこんな道を歩いた",
+		"何もない…良いことだ",
+	],
+	move_fail: [
+		"…壁か",
+		"通れない",
+		"別の道を",
+		"…行き止まり",
+		"戻る",
+		"…ここではない",
+		"迂回する",
+		"…無理だな",
+		"壁だ",
+		"…退く",
+		"道を変える",
+		"…仕方ない",
+		"別の方向へ",
+		"引き返す",
+		"…この体格では抜けられんか",
+	],
+	attack_miss: [
+		"…外した",
+		"甘い",
+		"次",
+		"…精度が足りない",
+		"もう一度",
+		"…まだだ",
+		"足りない",
+		"…修正する",
+		"当てる",
+		"…集中",
+		"甘かった",
+		"…次こそ",
+		"もう一撃",
+		"…戦場では許されん失態だ",
+		"外れた…だが",
+	],
+	combo_activated: [
+		"…畳みかける",
+		"今だ",
+		"続ける",
+		"隙あり",
+		"…連撃",
+		"…止めない",
+		"追撃",
+		"…ここで",
+		"攻める",
+		"…好機",
+		"逃さない",
+		"…押す",
+		"一気に",
+		"…歴戦の勘が告げている",
+		"手を緩めない",
+	],
+	enemy_defeated: [
+		"…終わりだ",
+		"排除完了",
+		"次",
+		"…片付いた",
+		"先へ",
+		"…倒した",
+		"終わった",
+		"…次の敵は",
+		"処理完了",
+		"…問題ない",
+		"片付けた",
+		"…先を急ぐ",
+		"一つ減った",
+		"…幾度目の勝利か、数えてもいない",
+		"次へ行く",
+	],
+	damage_taken: [
+		"…効くな",
+		"浅い",
+		"まだだ",
+		"…耐える",
+		"構わない",
+		"…痛みは知っている",
+		"この程度",
+		"…まだ立てる",
+		"問題ない",
+		"…致命傷ではない",
+		"耐えた",
+		"…平気だ",
+		"傷は浅い",
+		"…この体は幾多の傷を負ってきた",
+		"まだ戦える",
+	],
+	game_over: [
+		"…ここまでか",
+		"力及ばず",
+		"…無念",
+		"まだ…終われない",
+		"次は…負けない",
+		"…すまない",
+		"足りなかった",
+		"…認める",
+		"力が…",
+		"…これが限界か",
+		"不甲斐ない",
+		"…悔いはない",
+		"また、いつか",
+		"…力尽きた",
+		"…長い戦いだった",
+	],
+	trap_triggered: [
+		"…罠か",
+		"不覚",
+		"…やられた",
+		"浅はか",
+		"…注意が足りなかった",
+		"…失態",
+		"踏んだ",
+		"…気づかなかった",
+		"油断した",
+		"…迂闊",
+		"二度目はない",
+		"…侮った",
+		"甘かった",
+		"…戦場の罠はもっと巧妙だった",
+		"罠を踏むとは",
+	],
+	treasure_found: [
+		"…助かる",
+		"回復する",
+		"ありがたい",
+		"…使わせてもらう",
+		"丁度良い",
+		"…拾う",
+		"使える",
+		"…悪くない",
+		"これは良い",
+		"…必要なものだ",
+		"ありがたく",
+		"…助かった",
+		"いただく",
+		"…運が向いた",
+		"活用する",
+	],
+	rest_area_used: [
+		"…少し休む",
+		"回復",
+		"態勢を整える",
+		"…一息",
+		"ここで",
+		"…休息",
+		"体力を戻す",
+		"…しばし",
+		"休む",
+		"…助かる",
+		"体を休める",
+		"…古傷が疼く。少し休もう",
+		"少し、休む",
+		"…回復した",
+		"整えた",
+	],
+	floor_reached: [
+		"…新たな階層",
+		"進む",
+		"油断するな",
+		"…次の戦場",
+		"気を緩めるな",
+		"…降りる",
+		"新しい敵か",
+		"…構わない",
+		"どこまでも行く",
+		"…深くなる",
+		"先がある",
+		"…次の階だ",
+		"まだ続く",
+		"…昔もこんな深みに潜ったことがある",
+		"変わらず行く",
+	],
+	jump_success: [
+		"…飛ぶ",
+		"突破",
+		"一気に",
+		"…抜ける",
+		"行く",
+		"…跳んだ",
+		"着地",
+		"…問題ない",
+		"飛んだ",
+		"…それだけだ",
+		"越えた",
+		"…着いた",
+		"通過",
+		"…この体格で跳ぶのは得意ではないが",
+		"何でもない",
+	],
+	card_acquired: [
+		"…新しい手札",
+		"受け取る",
+		"使える",
+		"…悪くない",
+		"一枚追加",
+		"…確認した",
+		"手に入れた",
+		"…戦力になる",
+		"選ぶ",
+		"…これだ",
+		"加える",
+		"…必要なものだ",
+		"ありがたい",
+		"…デッキに入れる",
+		"判断する",
+	],
+	card_skipped: [
+		"…不要",
+		"見送る",
+		"今は要らない",
+		"…このままで",
+		"変えない",
+		"…十分だ",
+		"必要ない",
+		"…現状維持",
+		"見送った",
+		"…問題ない",
+		"変更なし",
+		"…今の手札で",
+		"不要だ",
+		"…パスする",
+		"このままでいい",
+	],
+};
+
+export const rareSpeechVariants: Partial<
+	Record<SpeechEventType, readonly string[]>
+> = {
+	move_success: ["…この景色、少し好きだ"],
+	enemy_defeated: ["…よくやった、俺"],
+	damage_taken: ["…正直、きつい"],
+	game_over: ["…もう少し、生きたかった"],
+	combo_activated: ["…悪くない、この感覚"],
+	treasure_found: ["…ありがたいな、本当に"],
+	trap_triggered: ["…くそっ"],
+	floor_reached: ["…少しだけ、わくわくする"],
+	card_acquired: ["…少し、嬉しいかもしれない"],
+	card_skipped: ["…次がある、きっと"],
+};
+
+export const sequenceVariants: Partial<
+	Record<`${SpeechEventType}_${SpeechEventType}`, readonly string[]>
+> = {
+	damage_taken_enemy_defeated: ["…借りは返した"],
+	move_fail_move_fail: ["…壁ばかりだ"],
+	trap_triggered_move_success: ["…二度はない"],
+	enemy_defeated_enemy_defeated: ["…続けて排除"],
+	attack_miss_enemy_defeated: ["…今度は当てた"],
+	damage_taken_move_success: ["…退く"],
+};
